+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # django-conman changelog
 
-## Upcoming release
+## 0.1.0 2017-09-28
 
 ### Backwards incompatible